Completed Shopping Cart/Order Page
This is the completed shopping cart/order now page. The "Order Now" button will move to a page which will also have a shopping cart on it. You can add and remove items from the cart as you please. The "Cart" button is placed on the upper right corner of every page so the user can quickly find their cart/add more items.